+++ eventsub-websocket/src/main/java/com/github/twitch4j/eventsub/socket/IEventSubSocket.java
@@ -1,4 +1,4 @@
-package com.github.twitch4j.socket;
+package com.github.twitch4j.eventsub.socket;
 
 import com.github.philippheuer.credentialmanager.domain.OAuth2Credential;
 import com.github.philippheuer.events4j.api.IEventManager;
